Updates backup and restore documentation
Using the newly created securedrop-admin "backup" and "restore" actions,
rather suggested manual activation of the virtualenv env. Bumped the
required version to 0.4, since the Admin tooling referenced here is
largely new.

Also noticed that the backup docs were incorrectly stating that the ATHS
info would be refreshed on the Admin Workstation. Although we should
definitely automate that update as the docs implied, we'll handle in a
follow-up issue. For now, simply instructing Admins to perform that
adjustment manually, then follow up with the `securedrop-admin
tailsconfig` to make the changes permanent.