Removes backup play from prod playbook
The prod playbook was overloaded to support backups and restores
if-and-only-if a complicated series of CLI flags were supplied. Now that
we have dedicated playbooks (and associated subcommands in the
securedrop-admin script), we no longer need the include in the main
playbook.

Also removed from the staging playbook, where it was off by default,
same as in the prod playbook.