Fixes 0.4-rc1 tagging
During cutting of the 0.4-rc1 tag for QA, we encountered a problem with
the update_version script, and had to hotfix it. A subsequent PR made
the change permanent, but in the process we bungled some of the
associated edits, and as a result wound up with files containing both
"0.3.12" and "0.4-rc1" for version strings.

We need to correct that prior to 0.4-rc2 (or 0.4, for that matter),
since the update_version script expects one-and-only-one "previous
version" to match. Therefore I'm manually reimplementing the 0.4-rc1
version string in a few places, then we can cut 0.4-rc2 and build.