tests: pages-layout: functional screenshots
A new directory is created for tests designed to navigate either the
source or the journalist interface and take screenshots. The navigation
will be done using the functions defined in tests/functional. However,
instead of focusing on asserting each navigation step follows the
expected sequence, the goal is to take a screenshot of a particular
state of the application.

The screenshots can be used to verify:

* the page rendering is visually correct
* the localized strings do not damage the page layout

It also helps increase the code coverage and will catch typos in template
expansions.