tests: pages-layout: repeat tests for each locale
The automatic webdriver fixture is implicitly added to each
test and does the following:

  * create a directory named after the currently active locale
  * set the browser to use the currently active locale

It is parameterized with each l10n found in the translations directory,
meaning tests are repeated for each locale. For instance:

TestSourceLayout::test_login[en_US]
TestSourceLayout::test_login[fr_FR]

etc. When the tests complete, the screenshot can be viewed in the
tests/pages-layout/screenshots/en_US etc.