Revert "i18n: add Flask-Babel requirement"
Change came in via #1924 but tests were not run (see #1826), so a
breaking regression due to the AppArmor profiles has been wrecking
builds in CI. We need to reland the requirements updates as part of the
i18n effort, but as a short-term fix, we're backing out the breaking
change from develop, pending a more durable and fully-tested
implementation.

Fixes #2070.

This reverts commit 10b41800d5524aef6c483c5c77d37f279e4fcecd.