Add test to reproduce overloaded Tor guard OSSEC alert
A Tor log event indicating that a Tor guard in use is overloaded
currently produces an OSSEC alert. While this alert is an
excellent candidate to be sent upstream to FPF for analysis,
there is no action that a SecureDrop administrator is expected
to take, making this a spurious OSSEC alert.

This test reproduces this spurious alert and is a regression test
for an OSSEC rule patch.