Creates Makefile target for HTML linting
Porting the Travis html-linting command to the top-level Makefile, so it's
easy to run HTML linting during local development. The HTML linter is a recent
addition to the developer environment, and it's been a bit finicky, so hopefully
empowering developers to run the linter locally will help us shake out problems.

Updates the developer docs to include mention of the new Makefile
target, rather than providing a one-liner for copy/pasting and
customization.