Add --page-layout option to turn off page-layout tests by default
The page layout tests can add significant time to the running of the
full test suite. However, we want to provide an easy option for the
developer to run them when checking the rendering of i10n strings
or when generating new screenshots for user guides.

This commit adds a --page-layout option in `conftest.py` and marks
each page-layout test with a pytest.mark. By default, the page-layout
tests will not run. However if the --page-layout option is run, all
tests will execute.