Removes unused custom reboot logic
The Ansible playbooks that configure the staging machines now reboot
automatically and conditionally, only on the first install, so we don't
need to force a reboot manually when testing CI anymore.

We can also remove the "reboot" role, since it was only used in one
place, and is now superseded by the reusable task list included in the
staging and prod playbooks.