Adds config tests for Apache logfiles
The OSSEC configuration was referencing nonexistent logfiles,
and we have strict rules about which Apache logfiles are permitted,
namely that no access or even error logs are written to disk for the
Source Interface.

In the staging environment, however, we permit source-error.log to aid
in debugging during development. The config test suite now reflects this
distinction.