Consolidates definedness checks for email validation
Moving the assert tasks for whether the email-related config vars are
defined to before the tasks that reference the vars. It's a good idea to
do this in the spirit of validation, even though the `sdconfig`
subcommand for the securedrop-admin script ensures that the vars are
defined.

More importantly, we need the definedness checks in separate tasks so
that we can convert the assert checks for the network calls into command
module invocations, so we can make use of retry/until.