tests: functional tests need more generous timeouts
The previous value of 5 (implicitly_wait or WebDriverWait) is known to
sometime be insufficient. Some travis ci runs failed because they
could not find an element, presumably because the the underlying
machine was too slow for the page to display.

Setting the value to 60 or more would surely cover even the slowest of
machine. However it also means that a test failing to find the desired
element in the DOM will report failure after 60 seconds which
significantly slows debuging.

The timeout was not always set and as a consequence the
screenshots generated with --page-layout from the pages-layout
directory ran without it, making them even more susceptible to fail
when run on slow machines.