tests: retry if connection to firefox fails
Although one would expect this is for selenium to do the needful,
there is a persistent race which leads to rare failures to connect. It
presumably happens when firefox is started, listening to the port but
not yet accepting connections. The simplest workaround is to retry.

There is ample research and trial / error when browsing
https://review.openstack.org/#/c/375258/ and the associated issues.