XFAIL functional tor connectivity test :(
This is a really annoying issue to debug and track down. I think its a needed
test but not critical to leave in CI right now because its been getting really
flaky (it works fine as soon as I login to CI manually and run it at terminal)
and the test case needs some refactoring and expansion. Preferably with a web-driver.
Happy to chuck this to another ticket.