Stop running tests as root in Travis
We've shaken out most of the pip dependencies woes, so we can finally
act like humans again and run the tests with normal user privileges.
Includes the `--page-layout` option and all tests are passing.

We also don't need to cd into the securedrop directory anymore, since
the bug in the test_manage suite has been resolved. Still skeptical that
the coverage logic is directory-agnostic, but will determine once a PR
is opened based on this branch.