Deletes legacy migration scripts
We had a number of migration scripts hanging around in the repo, left
over from the pre-0.3 days. SecureDrop 0.3 was released two and a half
years ago (the current release as of today is 0.4.2), so it's time to
say goodbye to these faithful helpers.

Most migrations these days are handled via preinst/postinst scripts in
the Debian package. If we need more comprehensive changes, we can write
them declaratively via Ansible and instruct Admins to run a dedicated
playbook.