isolate the srm function
When srm is delegated to a redis worker via rq, the modules it needs
will be imported in another python interpreter. To keep this to a
minimum, the srm function is isolated in a single file which only
depends on subprocess.

The former implementation was included in the store file which imports
gnupg. But gnupg runs processes and create directories when it is
imported. It will race with the journalist wsgi threads (each thread
has its own python interpreter) on a production server. And it will
also race with the journalist application run in a test environment,
inadvertendly re-creating the key directory after the teardown
function of the test removed the temporary directory in which it ran.

The redis server cannot be replaced by a thread run from the
journalist application because the production server has dozens of
them ran in each wsgi thread.