Creates custom yamllint config file
There are a 191 violations of the "line too long" rule if the max length
is 80 characters. That's too many to deal with at present, and it clutters up
the linting output, so let's relax the max length line rule to 200
characters, and preserve the "warning" status from the relaxed ruleset.

The yamllint man page states that yamllint will automatically detect and
apply a `.yamllint` file in the CWD, so it's technically unnecessary to
expand the fullpath in the Makefile, but doing so is more explicit for
contributors who may not be aware of the dotfile.