Lint YAML files with extensions .yml and .yaml
As of today, the project contains no files matching `^.*\.yaml`. That
may change in the future, however, and simply checking for `.yml` files
will cause linting to miss new files if they are added with the `.yaml`
extension. So, check for both. We can't use `file` because it's not YAML
aware.