Adds yamllint to develop pip dependencies
Ensuring that the Development VM can run the yamllint Makefile target.
The develop-requirements file is installed in the Dev VM, and also
mentioned in the dev docs as what developers should install within a
virtualenv on their workstations, in order to work on SecureDrop.

In the same stroke, we get yamllint available via pip in Travis, as
well, which is where the linting tasks run in CI.

Really, we should be sorting out the pip requirements files and
consolidating them responsibly. The need to do so is tracked in #2175.