Excludes certain shellcheck rules from linting
SC2001 [0] refers to use of sed for munging variables in Bash. The rule only triggers
on our update_version script, which admittedly is pretty difficult to read when
it comes to swapping out the periods within the version numbers (since
the periods are treated as special characters by sed).

SC2064 [1] refers to use of bash traps, and is violated only once in our
project, in the securedrop-keyring package preinst script. The usage
there works either way, so we could place an in-line ignore within the
file, but I figured that would only complicate review. Once we have
linting in place, we can continually revise to get up to par.

SC2181 [2] refers to exist code checking via `$?`, rather than in-line. This
is low priority, since it's stylistic more than functional, and besides
some of the files that showed violations here are slated for removal
(specifically the bash scripts for CI, which are being ported to
 Molecule).

[0] https://github.com/koalaman/shellcheck/wiki/SC2001
[1] https://github.com/koalaman/shellcheck/wiki/SC2064
[2] https://github.com/koalaman/shellcheck/wiki/SC2181