Makes find command for yamllint cross-platform
The `-regextype=posix-extended` is not cross-platform; specifically, it
doesn't work on macOS. Let's use find to print all paths (except those
pruned), then filter for YAML files with grep. Using the `-E` flag to
grep for "extended" regex, which is equivalent to the missing find
option. (Reminder: do one thing and do it well!)

Also prefixing the find command with `@` to prevent the command from
being printed to stdout, since it's now garish.