tests: PAGE_LAYOUT_LOCALES controls --page-layout scope
By default the current implementation makes screenshots for every
locale found in the securedrop/translation directory. This is
impractical and never actually used because it takes too long. Instead
create screenshots for all locales listed in the coma separated string
from the PAGE_LAYOUT_LOCALES which defaults to en_US.