Provides "." find target to shellcheck linter
We can't use "$(PWD)" as we do elsewhere in the Makefile, since the
absolute path on the developer workstation won't match the temporary
mount path on the container. Omitting the directory to search from the
find command works well in Linux (tested Debian 9), but not with macOS's
BSD variant, as reported by @redshiftzero during PR review.