Add unit test for /create on source interface multiple times
If a source accesses the /create route twice, they will make two
attempts to insert a new source into the database. Only the first
should be successful. This commit adds a unit test for the case
of subsequent visits to /create, ensuring that duplicate codenames
are not inserted, an event is logged, and strange errors or logs do not
appear to the user.

This unit test actually reveals a bug: the db_session was not rolled
back in the existing exception handling code, so a 500 will be thrown.