Builds securedrop-config package along with other debs
Reusing the generic-pkg role to build the new `securedrop-config`
package along with the others.

Sets securedrop-config as dependency of securedrop-keyring

Adds local pkg vars for securedrop-keyring. Since `securedrop-keyring`
depends on `securedrop-config`, the latter must be installed first, so
the order of items the local packages list var is important.

Runs the usual collection of tests over the new deb package, as well.

Includes entry in the `update_version.sh` script so that the config
package will have updates made to it, as well.