Add testing for generate_unique_codename's long codename flow
The code-under-test here is just a guard against future changes, but
this unit test verifies that a codename greater than the maximum
codename length cannot be generated.

Note here that in the mock object for genrandomid() here I used a
side_effect that is an iterable such that eventually a valid codename
is produced. If a valid codename is never presented in
generate_unique_codename, it would loop forever.