Add functional tests for session timeout on source interface
We add an optional keyword argument `session_expiration` to the
test setup to be used for tests that require the session
timeout to be short.

The timeout cannot be -1 and instead must be a more realistic value
(here 1-2 seconds) as /create will fail if 'codename' is not stored
in the session.