Improve the login form usability by making input fields larger
In this commit, we are changing the login form style so the input fields
are larger, specially in big screens.

The inputs are set to a maximum width of 445px, the ideal size in order
to display all the characters of the password if the user is using
diceware passphrases.

The login form is now also mobile friendly, since it has a responsive
width, and the input fields will adapt to the screen size as well.

This commit fixes #2288