Write improved test for col_process on nonexistant action
The original unit test for col_process is mocked to an
unnecessary degree. The behavior it is trying to test for is also
incorrect: col_process should return a 500, not a 404, when the
action selected is invalid.

The test is replaced with one that makes minimal use of mocks,
and verifies that a 500 is thrown when an invalid action is
selected.