Creates securedrop-admin subcommand for fetching logs
Uses a single entrypoint in the top-level "securedrop-admin" script to
provide a convenient method for fetching, compressing, and encrypting
the logs for sending to FPF. We've often discussed the need for a
convenient action to prepare logs for support purposes, and 0.4.4
provides the perfect opportunity to ship the feature.

The GPG fingerprint used for encrypted the tarball is not customizable;
future versions of the script should provide override capability via
`--recipient` or similar.