Patches Ansible 2.2.3 -> 2.3.2
Updates the Ansible version to 2.3.2, in order to address CVE-2017-7481 [0],
which was patched in 2.3.1 [1].

Updates the Ansible version throughout the project, including both
the hash-pinned requirements file for SecureDrop Administrators, and the
sundry requirements files used by developers and in CI for testing. We
have a sanity check on forcing a supported Ansible version via a
callback plugin, so that hardcoded version check has been incremented,
as well.

[0] https://access.redhat.com/security/cve/CVE-2017-7481
[1] https://github.com/ansible/ansible/blob/70c58e74ab492ea17df893bb4692f3b64a755858/CHANGELOG.md#231-ramble-on---2017-06-01