Add SecureDrop backup and restore roles
The source-restore role puts the sources back on the server by
inserting them into the fresh database, adding their keys into the
keyring, and creates the source folder.

This creates a backup script that is copied to the application server
and ran to validate the source rows contain the expected data. It then
exports the secret keys only of the sources and saves everything as a
JSON file.

At the end of each role is a task that ensures the database has the
correct ownership. If the database is not owned by www-data, then
the database will be read-only.

Signed-off-by: Michael Sheinberg <m.sheiny@gmail.com>