testinfra: do not be strict about IPv6 tests failing
https://github.com/freedomofpress/securedrop/commit/3da5605
temporarily expects IPv6 tests to fail in CircleCI because the virtual
machine does not reboot. However, when run on a laptop they will
succeed because IPv6 is presumably already properly set, even if a
reboot did not happen.

By removing the strict=True we accept the test may fail or succeed,
depending on the context.