tests: wait up to one minute for the worker to srm
It is not an error for a machine to require more than 5 seconds to srm
a file. What we really want here is for the function to not wait
forever in case nothing is happening. Since all files involved in
tests are small, it is reasonable to assume that they can be removed
within 60 seconds, even on the slowest virtual machine running on the
slowest hardware.