Remove fork of synchronize.py module
Back in May (d803bc7b8c4), when we were still under ansible 2.2.x we had to pull in
the synchronize module from upstream in order to get rsync playing nice
with docker. Now that we are on 2.3+ there is some piece of code we are
missing from upstream that is causing problems. Lets drop our fork and
rely on upstream synchronize module now :)