Updates pip-tools 1.9.0 -> 1.10.1
Provides updates to the `--generate-hashes` option, which is required for
pinning based on checksum for the Admin Workstation requirements.
Ensures that checksums for all valid wheels are included, not just the
current platform. Should help maintain compatibility across developer
workstations, but also allow developers to update the Tails-specific
requirements files reliably. Includes a number of bugfixes to the
hashing logic, as well [0].

We pin checksums on the Admin Workstation because the Python
dependencies are pulled in via pip, and therefore do not afford GPG
signature verification, as we have with the git repo and apt packages.

We should really be pinning based on hash even for the development
environments, as well.

[0] https://github.com/jazzband/pip-tools/releases/tag/1.10.0