Updates test-requirements files
Technically a no-op: we already had strong version-pinning, and this
change just adds an updated header to the top of the file, automatically
generated by pip-compile.

We are *not* pinning checksums, since the test requirements files are
installed on the staging VMs, which run a very old version of pip
(1.5.4, rather than 8+ required for reading hashes in requirements).