l10n: add SUPPORTED_LOCALES to sdconfig
Allow the sysadmin installing (or updating) SecureDrop to specify the
content of the SUPPORTED_LOCALES variable when running

   ./securedrop-admin sdconfig

The updated value from the site-specific file will override the value
in config.py in the same way DEFAULT_LOCALE is overriden.

The list of locales entered via sdconfig is validated to match
existing directories from the securedrop/translations directory.