source: fix regression on /generate page display
22de405cb78d2a2cb9bb7c6eff55be4c55afa68c improved the display of the
/generate page on mobile. Unfortunately it also introduced a
regression when the screen is large enough to accomodate the
.attachement and .message blocks next to each other: the message block
only occupies half the space. The user would be confused and click on
the right size of the block, doing nothing instead of focusing on the
text area.

Conditionally activate the mobile display if the viewport has a size
of 768px maximum. This is consistent with a similar adaptation
introduced for the index page in _source_index.sass file