Adds Makefile target for updating app-code pip reqs
Overlooked during the creation of the new `update-pip-dependencies`
Makefile target, and pointed out by @redshiftzero during review. We do
need the Makefile target to update *all* project dependencies, and the
line for the app-code requirements update via pip-compile is added via
this commit.

However, I have not actually updated the app-code requirements, since I
see that as a separate test (and will surely require updates to the
config tests, which hardcode many pip versions so we don't inadvertently
break things).