Ensure admins able to write to HTTPS validation dir
When under-going HTTPS validation with a CA, this will involve dropping
arbitrary files in order to prove ownership over the domain. We discuss
this language in the documentation, but in reality, apparmor will block
attempts to create and add files to this folder. This commit adds
an empty directory and the matching apparmor and apache configuration
for usage for this purpose.