app: use lazy_gettext instead of gettext in wtforms
The wtforms classes are evaluated outside of the context of a
request. When gettext is used to set a parameter, it has no way to
know which language is going to be used and default to english.

We need to use lazy_gettext which delays the evaluation until the
string is used, i.e. when the user session is active and the language
preferences can influence the translation.