Fix #2656: Ensure buttons with lots of text go to two lines gracefully
The buttons on the generate page on the source interface have long
text in a few languages. We should have them gracefully go to two
lines if needed, with space between them and they should be right
aligned.

The space is done via a new class `btn--space` (following BEM [0])
which specifies a 10px margin so that this renders nicely.

Then `pull-right` is applied to each button individually so the
buttons (not just the div) individually are right aligned (in RTL
languages they will be left-aligned).

[0] https://css-tricks.com/bem-101/