Adds securedrop-metapackage file tree
Using flat files in a directory tree to provide the skeleton for the
`securedrop-grsec` package. Doing so will enable us to reuse the
existing "build-generic-pkg" role logic in preparing the deb package.

All files are hardcoded, so we'll need to adjust the versions carefully.

**IMPORTANT:**

The "Version" field in the Debian control file is set to ``3.14.79+r1`
since that's what's currently in production as of 2017-12-07. Future
versions of the SecureDrop metapackage must increment from that value,
otherwise updates will not be applied.