Builds securedrop-grsec metapackage
Along with the other `securedrop-*` packages, let's build
`securedrop-grsec` as part of the normal release cycle. Still out of
band, in another repository, is the underlying grsecurity-patched kernel
image package that is referenced by the metapackage.

The metapackage also provides helper scripts such as ensuring PaX flags
are set correctly on the grub binaries, so that the `update-grub`
command succeeds when upgrading between grsecurity-patched kernels.

The sysctl settings left over from the Stack Clash mitigations can be
moved to the `securedrop-config` package once that ships.