Provides baseline config tests for securedrop-grsec
Tacks on a testinfra test to perform rather superficial testing of the
`securedrop-grsec` metapackage. Piggybacks on the testinfra fixtures
already in place in the `test_securedrop_deb_package.py` test suite.

Does *not* validate that the hardcoded file contents are correctly
structured; simply ensures the files have been copied in and exist in
the built package. The files are crucial for ensuring kernel image
updates are applied successfully.