source: resolve generate interaction confusion
If sources arrive at the generate page, it is because they clicked on
the "Submit Document" button in the index page. We remove the "USE
EXISTING CODENAME" from the generate page because it means the exact
opposite. If sources change their mind, they can go back to the index
page.

Removing this button is important for another reason: the wording may
confuse sources. Some may think "EXISTING CODENAME" refers to the
codename displayed in the generate page although it means the exact
opposite. And the "NEW CODENAME" could also be interpreted as a way to
create a codename that is different from the one displayed in the
page (for instance because it is difficult to memorize). But it
actually means the opposite.

Removing this button is also a way to avoid confusion in the
future. The previous wording of the buttons (using CONTINUE) was
slightly confusing. And I merged the current wording because it
sounded like it was an improvement but it turns out to be worse.  This
is a hint that the problem is not in the wording of the buttons.