Breaks out postfix config into discrete role
Previously we configured both postfix and procmail inside the
"ossec-server" role, which was essentially shoehorning the config. It's
appropriate to target the "securedrop_monitor_server" group with those
config items, but technically there's not literally OSSEC server, but
rather a separate service that deserves its own configuration logic.

Tidies up some of the vars. We've documented use of the
`ossec_from_address` publicly, so we can't simply drop reference to it.
Since the associated logic now resides in the "postfix" role, it should
be prefixed with the `postfix_` namespace. That's done, old values of
`ossec_from_address` set at the site level, if set, will be honored.

(cherry picked from commit 6aaf88a536633a3c8678b933197de01df0962f00)