Add temporary iptables to the top of the chain
Without this explicitly set, ansible will append the iptables to the
bottom of the chain. Since we have some explicit DROP lines at the
bottom of the INPUT chain, this means that the exception gets added
AFTER a DROP line. Meaning... that the line basically never got
evaluated.

This commit resolves that issue by always inserting at the top of the
chain (both INPUT and OUTPUT). There is a following clean-up task that
will remove this afterwards as usual.