Adds config tests for Referrer policy headers
Tracking the config change in the testinfra test suite in order to guard
against regressions. The tests now expect the header to be present in
the Apache config for both the Source and Journalist Interfaces. The
tests do *not* confirm that the header is set on user-agents, but I
confirmed that via manual browser testing during review.