Add task to inspect if ntpd is already running
We should inspect the stderr of the ntpdate task. If it fails due to the NTP socket in use,
we can ignore it and proceed, since that means that ntpd was setup already and the system
time is thus accurate. However, if it fails for some other reason, then we want to fail
out at these tasks and not proceed further in the install so the admin can debug.