source: use SUBMIT DOCUMENTS in generate
The generate page explains the importance of the code name to the
sources, after they have clicked on the SUBMIT DOCUMENTS button in the
index page. It is an intermediate page that only exists for
information purposes.

To move forward from the generate page and actually submit documents,
it makes sense to re-use the same wording as the index page,
i.e. SUBMIT DOCUMENTS instead of USE NEW CODENAME which could be
confusing.