tests: activate journalist session expiration verification
The test did not verify session expiration because the session was not
set in the first place. The VALID_PASSWORD constant is not the
password and therefore the session is not established because the
login fails. This is semi-hiddent by the status_code 200 assert which
is always true after a login: the only difference is in the flashed
message.

A secondary problem is that config.SESSION_EXPIRATION_MINUTES does not
influence the next request. When the next request runs, it sets
session['expires']. Only the second request will expire after noticing
the session['expires'] is in the past.

An assert on session['uid'] is added to prevent further regressions