journalist: redirect to main.login on password change
Similar to setup_g, if the session expires because the password
changed, go to the login page instead of staying on the edit
page. This does not change the user experience because they will be
redirected to the login page anyway, but it makes things more explicit
for the developer and easier to test.