circleci: do not cache docker layers
When a RUN apt-get update layer is cached it may become obsolete. This
will manifest with apt-get install failing on 404 because the cache
references packages that no longer exist.

The developer who sees this error is unlikely to link that to the
fact that docker layers are cached accross runs. And it will not occur
to them to change the line of the Dockerfile to force an update.

The benefit of maintaining a cache is to speed up tests (skipping
apt-get install and pip install on every run). But this benefit is
less appealing when developer are confused by the errors it may
induce.