i18n: move format() outside of ngettext
Although it makes more sense to write:

ngettext('1 doc', '{doc_num} docs'.format(doc_num=docs), docs)

the pybabel parser fails to recoginize the call and does not include
it in messages.pot. We add the .format() outside instead. It looks
strange because '1 doc' does not need it but ... it does not hurt.