source: no need to be so specific about tracking
The codename can be seen as a way of 'tracking', but SecureDrop does
its very best to not track anything about the sources. Saying "we
don't do the traditional tracking" implies for the suspicious of mind:
"so we fall back on even more sophisticated ways of tracking."

Revert to the simple form "we do not track users".