tests: wait for source key generation in functional tests
All functional tests assume the source GPG key is created when a
document or a message is submitted. But it really is generated
asynchronously and therefore races with the tests. It is fine most of
the time but leads to subtle bugs.

For instance when a journalist functional test tries to reply to a
source. If the GPG is being created the selector will not be
displayed: instead the "flagged for reply" display will show and the
test will fail to reply.

This race was incorrectly fixed by waiting for the key to be generated
before deleting it, because it happened relatively frequently.

A support function is added to wait until the GPG key for a given
source becomes available. It is called whenever a source submits a
document and effectively waits until the async_genkey completes.