app: always chmod u+w the logo in postinst
Prior to 0.5.1 a custom logo was installed with u-w. Now that the
journalist app can write a new logo, this is a problem and write
permissions must be restored.

We chmod u+w unconditionaly because it is cheap although it will be
useless for upgrades after 0.5.1. And it will help people upgrading
from a version < 0.5.1 directly to a version > 0.5.1.

Also get rid of a test that prepared a future custom logo test and
became obsolete.