Remove regular staging-test in bid to keep complexity down
This is a hot fix to reduce the number of AWS calls we are making per CI
run. Specfically, we have started seeing issues with rate-limiting
against the `CreateKeyPair` that need to be mitigated. Really a longer
term goal is to rethink how/when we want to run the staging full VM
tests in CI. Shorter term we can also analyze how we create SSH keys -
maybe we take the same key and shove it in circleci?

(cherry picked from commit 5ecddcf294f997b1d641f714a8606f0f674b38d7)