Adds Makefile target "ansible-config-lint"
Target name is intentionally a bit verbose in order to discrimate from
"ansible-lint", which is a dedicated tool for linting Ansible playbooks
and the roles called therefrom, ensuring adherence to best practices.
That's a separate tool, and one we use elsewhere, whereas the new test
suite being introduced ensures SecureDrop-specific config choices
that we've added in this repository.

Hooked up the new test to the general "make lint" target, as well. The
new tests are *fast*, so no concerns with CI delays.