Implements max_fail_percentage in all playbooks
The primary concern is for the "securedrop-prod.yml" playbook, which is
what Admins use to configure a production SecureDrop instance. In that
playbook, any error encountered during provisioning should cause an
immediate failure, aborting execution and reporting an error message for
review.

If the max_fail_percentage is *not* set to 0, debugging errors becomes
challenging, often involving lots of scrollback to piece together what
specifically went wrong.

In order to provide a sane strategy for testing adherence to this
requirement, we'll test all playbooks (defined as "YAML file in
ansible-base directory, and not the old prod-specific.yml vars file")
and ensure the option is set on each and every play. That's a bit
aggressive, but the tests are extremely fast, and will likely pay off
as we bring new contributors into the project who may not be familiar
with this rather esoteric Ansible option.

(cherry picked from commit 1421b404820901c3d76b0311d9033f1499d759df)