Removes unused "upgrade" playbook
Also removes associated "upgrade" role, which was designed for the
migration to SecureDrop 0.3, released on 2015-02-12. These files
have not been used since then, and so should be removed from the
repository.

As a side note, these files were first written, and last touched, by the
late James Dolan. RIP, James. I cannot express what a profound pleasure
it was to work with you. You are missed, dearly.