Reenables OSSEC connectivity config test
Partial reversion of 3da5605ae7, in which several config tests were
intentionally disabled for CI optimization. Now that we've resolved the
unreliable OSSEC registration flow, the strict "xfail" on the OSSEC
connectivity test was reporting failure: because it was unexpectedly
passing!

Renabled the OSSEC test, but left the other changes from 3da5605ae7
in place. We'll get to those in their own time.