securedrop-admin: disable DNS checks
In 7c54d517efa2aa22b7ff34b6f5903c9e6918e7a2 it was decided to no
longer check DNS entries because we could not make it work reliably in
tails. Disable in the securedrop-admin re-implementation as well but
keep the implementation so it can be re-activated and fixed instead of
being re-implemented from scratch.