Lint all Python files with flake8
The Makefile "flake8" target is now more comprehensive, using a
blacklist approach rather than a whitelist approach to decide which
files to lint. This strategy should catch new Python files introduced to
the repository, so that CI will lint them even if they're in new
directories.

Rather than using a find-based approach, as with the "shellcheck" target,
we'll run flake8 on the entire repository and trust flake8 to figure out
which files it wants to inspect.