Flake8: Lints 0.3_collect.py backup script
Minimal usage on this script, but let's clean it up anyway. If it's in
the repo and a Python file that we control, it should pass flake8
linting.

Added a few "noqa" lines here to permit non-standard imports, which are
required due to the use of "config.py" for SecureDrop configuration.

Also tacking on changes to the "restore.py" script, which is closely
related.