Flake8: Linting profile_tasks Ansible callback
The "profile_tasks" Ansible callback plugin was added to the repository
back in the Ansible 1.x days. It's no longer necessary, since as of
Ansible v2 task profiling was merged into core, and can be enabled via a
whitelist option for callbacks in ansible.cfg. Removing the plugin
should be handled separately, since all the changes in the current push
involve strictly flake8 linting.