update-user-guides: refactor to run as non-root
When running update-user-guides it creates files that are on the host,
to update screenshots. The script runs as root in the docker
container and these files must be chown to developer user to avoid
permission problems. But a number of other files can also be created
as root as a side effect (pyc, etc.) and ensuring they all belong to
the user is error prone.

Instead of running the script as root, run it as the develper user and
rely on sudo for actions that require root permissions in the
container.

A VNC proxy is exposed to enable the developer to run

  xtightvncviewer localhost:1

and see the X11 display used by the tests, for debugging purposes.