docs: add instructions to developers to show the l10n menu
When setting up the development virtual machine, the language menu
will not be displayed by default because the .po files are not
compiled into .mo files. We add instructions to setup the development
virtual machines so the menu is always displayed.

The translations need to be updated manually by running this
compilation step every time a source string changes in the source or
journalist interface.