Fixes Vagrant box downloading for Vagrant 1.x
The logic to fetch and update Vagrant boxes from Hashicorp's hosting
broken suddenly on or around 2018-02-04. Looks like backend changes on
their part are the culprit, and newer versions such as Vagrant 2.0.2 may
not be affected.

Let's use a workaround, which should be temporary, assuming appropriate
redirects are implemented on the backend, in order to keep fresh
spin-ups of the dev env, as well as updates to the base boxes for
regular developers, working as expected.

Closes #2970.